redo api call classes namespacing
